feat (47): Updating unit tests for the Favorite movies list screen
